fix linking of XML objects
