+++ src/gsMaterialMatrixContainer_.cpp
@@ -7,5 +7,9 @@ namespace gismo
 {
   CLASS_TEMPLATE_INST gsMaterialMatrixContainer<real_t>;
 
+namespace internal
+{
+    CLASS_TEMPLATE_INST gsXml< gsMaterialMatrixContainer<real_t>>;
+}
 }
